Removed outcomes/constants/trade-types.js moved all references to BUY and SELL to transactions/constants/types.js
