+++ src/modules/outcomes/components/outcome-trade.jsx
@@ -9,7 +9,7 @@ import ComponentNav from 'modules/common/components/component-nav';
 import EmDash from 'modules/common/components/em-dash';
 
 import { SHARE, MICRO_SHARE, MILLI_SHARE } from 'modules/market/constants/share-denominations';
-import { BUY } from 'modules/outcomes/constants/trade-types';
+import { BUY } from 'modules/transactions/constants/types';
 import { BIDS, ASKS } from 'modules/order-book/constants/order-book-order-types';
 import { SCALAR } from 'modules/markets/constants/market-types';
