--- src/modules/outcomes/constants/trade-types.js
@@ -1,2 +1,0 @@
-export const BUY = 'buy';
-export const SELL = 'sell';
